requirements.txt
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
Pillow>=10.0.0
mss>=9.0.1
opencv-python>=4.9.0
numpy>=1.26.0
